tests/wpt/web-platform-tests/app-history/currentchange-event/currentchange-app-history-updateCurrent.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(t => {
  let oncurrentchange_count = 0;
  appHistory.oncurrentchange = t.step_func(e => {
    oncurrentchange_count++;
    assert_equals(e.from, appHistory.current);
    assert_equals(e.navigationType, null);
    assert_equals(appHistory.current.getState(), "newState");
  });
  appHistory.updateCurrent({ state: "newState" });
  assert_equals(oncurrentchange_count, 1);

  // "Updating" the state to the current state should still fire currentchange.
  appHistory.updateCurrent({ state: appHistory.current.getState() });
  assert_equals(oncurrentchange_count, 2);
}, "AppHistoryCurrentChangeEvent fires for appHistory.updateCurrent()");
</script>
